[!3][PADDING] Add encoder padding checker
# Why is the change needed?

A generic aspect we want to test is the proper handling of padding.
To this aim, we want to provide a base test case that ensures that the output
of an encoder module behaves properly with respect to padding.

# What changes does the patch introduce?

Adds a new package that contains a module wrapper and a base test class to
be extended to test the proper handling of padding.

# How was this patch tested?

Added UTs