[!5][AUTOREGRESSIVE] Add tester for causal modules (e.g., NMT autoregressive decoder)
# Why is the change needed?

In attention-based encoder-decoder architectures (e.g. Transformer), the decoder must not attend to the future when predicting the output, otherwise the training would be trivial and inference impossible.
This property, sometimes referred as causality, has to be enforced on every module in autoregressive decoders.

# What changes does the patch introduce?

Adds tester for autoregressive causality.

# How was this patch tested?

added UTs